pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "maafw-cli"
version = "0.1.0"
description = "A command-line interface for MaaFramework"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "MistEO", email = "misteo@outlook.com" }
]
keywords = [
    "maa",
    "maaframework",
    "automation",
    "android",
    "adb",
    "ocr",
    "cli",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "maafw>=5.2.6",
    "click>=8.0.0",
    "opencv-python>=4.0.0",
    "platformdirs>=4.0.0",
]

[project.scripts]
maafw-cli = "maafw_cli.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/maafw_cli"]

[tool.hatch.build.targets.sdist]
include = [
    "src/maafw_cli/**/*.py",
    "README.md",
    "LICENSE",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
